Add override 'factoryish' method.
This basically allows users to build configuration namespaces, either in
environment variables or in their folder structure.

This is great for having multiple services utilize the same base
configuration but override some individual pieces of configuration
specific to that service.

For the environment variable loader, the user can use a different prefix,
like MYAPP_ and MYAPP_BACKEND_ to select configuration.

For a loader like zookeeper or whatever, the user could change the
zookeeper path root, and for files the user could provide a small dropin
file specifiying just the additional configuration.

How all this gets configured is another story, but we'll leave that up
to the users imagination