Fix ups
Fix dict loader so it can load additional properties - you passed them
in so you totally want those, right?

Fixed issue where schema would be modified when you load data. This was
a weird one took an hour or so to debug. Basically, the
DefaultSettingValidator will entangle the schema and config such that
updating the config object will update defaults in the schema. We don't
want that, since we want the schema to never ever ever change. So, we
freeze it when we get it, and only pass jsonschema a thawed copy of it.

Finally, somehow we managed to miss having boolean conversion - kind of
important and probably want to improve this later to do a yaml bool kind
of casting.